job_agent/db/connection.py
import os
import psycopg2
from psycopg2.extras import RealDictCursor
from dotenv import load_dotenv

load_dotenv()

def get_conn():
    return psycopg2.connect(os.environ["DATABASE_URL"])

def get_cursor(conn):
    return conn.cursor(cursor_factory=RealDictCursor)
