requirements.txt
anthropic
psycopg2-binary
pgvector
python-dotenv
jinja2
weasyprint
voyageai
